additional unit test, rename factories/courses.js to factories/resources.js
