+++ static/js/factories/resources.js
