+++ static/js/reducers/courses_test.js
@@ -4,7 +4,7 @@ import sinon from "sinon"
 
 import IntegrationTestHelper from "../util/integration_test_helper"
 import { actions } from "../actions"
-import { makeCourse } from "../factories/courses"
+import { makeCourse } from "../factories/resources"
 
 describe("course reducers", () => {
   let helper, response, dispatchThen
