+++ static/js/util/integration_test_helper.js
@@ -20,6 +20,7 @@ import * as postAPI from "../lib/api/posts"
 import * as commentAPI from "../lib/api/comments"
 import * as livestreamAPI from "../lib/api/livestream"
 import * as courseAPI from "../lib/api/courses"
+import * as bootcampAPI from "../lib/api/bootcamps"
 import * as widgetAPI from "../lib/api/widgets"
 import rootReducer from "../reducers"
 import * as utilFuncs from "../lib/util"
@@ -56,6 +57,7 @@ export default class IntegrationTestHelper {
       commentAPI,
       livestreamAPI,
       courseAPI,
+      bootcampAPI,
       widgetAPI
     ].forEach(apiModule => {
       for (const methodName in apiModule) {
